requirements.txt
fpdf==1.7.2
olefile==0.44
Pillow==4.1.1
PyPDF2==1.26.0
